+++ drools-model/drools-model-compiler/src/test/resources/logback-test.xml
@@ -10,7 +10,7 @@
   <logger name="org.kie" level="info"/>
   <logger name="org.drools.compiler.kie.builder.impl" level="ERROR" />
   <!--<logger name="org.drools.modelcompiler.builder.generator.expressiontyper" level="debug"/>-->
-  <logger name="org.drools.modelcompiler.builder" level="debug"/>
+  <!--<logger name="org.drools.modelcompiler.builder" level="debug"/>-->
   <!--<logger name="org.drools.compiler.reteoo.compiled" level="debug"/>-->
 
   <logger name="org.drools" level="info"/>
